Fixing autoload.
The autoload settings / namespace structure / folder structure are not correct. Composer cannot autoload this package without changes.

Since the folder path is `src/Bhutanio/BEncode` the namespace for a psr-0 package has to be `Bhutanio\BEncode` and the full class name has to be `Bhutanio\BEncode\BEncode`.

If this was not your intention then perhaps you should remove the `BEncode` folder and place `BEncode.php` directly into `src/Bhutanio` which would also fix the problem.